haskel-user-service/pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
